tools/animation/requirements.txt
matplotlib==3.9.2
numpy==1.23.3
paho_mqtt==2.1.0
pyserial==3.5
